backend/requirements.txt
fastapi==0.124.4
uvicorn
starlette
pymongo==4.6.1
motor
redis==5.0.1
pydantic==2.5.3
python-multipart==0.0.6
bcrypt==4.1.2
python-jose==3.3.0
passlib==1.7.4
email-validator==2.1.0.post1
chromadb
pypdf
sentence-transformers
torch
transformers
aiofiles
Pillow
pydantic-settings
